third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/suborigins/crossorigin/suborigin-cors-eventsource-expected.txt

CONSOLE ERROR: EventSource cannot load http://127.0.0.1:8000/eventsource/resources/es-cors-basic.php?. The 'Access-Control-Allow-Suborigin' header has a value '' that is not equal to the supplied suborigin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: EventSource cannot load http://127.0.0.1:8000/eventsource/resources/es-cors-credentials.php?. The 'Access-Control-Allow-Suborigin' header has a value '' that is not equal to the supplied suborigin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: EventSource cannot load http://127.0.0.1:8000/eventsource/resources/es-cors-basic.php?. The 'Access-Control-Allow-Suborigin' header has a value '' that is not equal to the supplied suborigin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS EventSource: anonymous, ACAO: http://127.0.0.1:8000 
PASS EventSource: anonymous, ACAO: * 
PASS EventSource: use-credentials, ACAO: http://127.0.0.1:8000 
PASS EventSource: anonymous, CORS-ineligible resource 
Harness: the test ran to completion.
